pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.0.0",
    "numpy>=2.3.5",
    "pandas>=1.5",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "typer>=0.20.0",
]
